+++ sdk/keyvault/keyvault-keys/src/transformations.ts
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { DeletedKeyBundle, KeyBundle } from "./generated/models";
+import { DeletedKeyBundle, KeyAttributes, KeyBundle, KeyItem } from "./generated/models";
 import { parseKeyVaultKeyId } from "./identifier";
-import { DeletedKey, KeyVaultKey, JsonWebKey, KeyOperation } from "./keysModels";
+import { DeletedKey, KeyVaultKey, JsonWebKey, KeyOperation, KeyProperties } from "./keysModels";
 
 /**
  * @internal
@@ -17,7 +17,7 @@ export function getKeyFromKeyBundle(
 
   const parsedId = parseKeyVaultKeyId(keyBundle.key!.kid!);
 
-  const attributes: any = keyBundle.attributes || {};
+  const attributes: KeyAttributes = keyBundle.attributes || {};
   delete keyBundle.attributes;
 
   const resultObject: KeyVaultKey | DeletedKey = {
@@ -50,6 +50,73 @@ export function getKeyFromKeyBundle(
     (resultObject as any).properties.scheduledPurgeDate = deletedKeyBundle.scheduledPurgeDate;
     (resultObject as any).properties.deletedOn = deletedKeyBundle.deletedDate;
   }
+
+  return resultObject;
+}
+
+/**
+ * @internal
+ * Shapes the exposed {@link DeletedKey} based on a received KeyItem.
+ */
+export function getDeletedKeyFromKeyItem(keyItem: KeyItem): DeletedKey {
+  const parsedId = parseKeyVaultKeyId(keyItem.kid!);
+  const attributes = keyItem.attributes || {};
+  const deletedBundle = keyItem as DeletedKeyBundle;
+
+  const abstractProperties: KeyProperties = {
+    createdOn: attributes?.created,
+    enabled: attributes?.enabled,
+    expiresOn: attributes?.expires,
+    id: keyItem.kid,
+    managed: keyItem.managed,
+    name: parsedId.name,
+    notBefore: attributes?.notBefore,
+    recoverableDays: attributes?.recoverableDays,
+    recoveryLevel: attributes?.recoveryLevel,
+    tags: keyItem.tags,
+    updatedOn: attributes.updated,
+    vaultUrl: parsedId.vaultUrl,
+    version: parsedId.version
+  };
+
+  return {
+    key: keyItem,
+    id: keyItem.kid,
+    name: abstractProperties.name,
+    keyOperations: deletedBundle.key?.keyOps,
+    keyType: deletedBundle.key?.kty,
+    properties: {
+      ...abstractProperties,
+      recoveryId: deletedBundle.recoveryId,
+      scheduledPurgeDate: deletedBundle.scheduledPurgeDate,
+      deletedOn: deletedBundle?.deletedDate
+    }
+  };
+}
+
+/**
+ * @internal
+ * Shapes the exposed {@link KeyProperties} based on a received KeyItem.
+ */
+export function getKeyPropertiesFromKeyItem(keyItem: KeyItem): KeyProperties {
+  const parsedId = parseKeyVaultKeyId(keyItem.kid!);
+  const attributes = keyItem.attributes || {};
+
+  const resultObject: KeyProperties = {
+    createdOn: attributes.created,
+    enabled: attributes?.enabled,
+    expiresOn: attributes?.expires,
+    id: keyItem.kid,
+    managed: keyItem.managed,
+    name: parsedId.name,
+    notBefore: attributes?.notBefore,
+    recoverableDays: attributes?.recoverableDays,
+    recoveryLevel: attributes?.recoveryLevel,
+    tags: keyItem.tags,
+    updatedOn: attributes.updated,
+    vaultUrl: parsedId.vaultUrl,
+    version: parsedId.version
+  };
 
   return resultObject;
 }
